refactor io code to prevent circular import
