+++ nimare/io.py
@@ -14,6 +14,8 @@ from scipy import sparse
 
 from nimare.dataset import Dataset
 from nimare.extract.utils import _get_dataset_dir
+from nimare.utils import nimads_to_dataset
+from nimare.nimads import Studyset, Annotation
 
 LGR = logging.getLogger(__name__)
 
@@ -31,43 +33,21 @@ def _create_name(resource):
     return "_".join(resource.name.split()) if resource.name else resource.id
 
 
-def convert_nimads_to_dataset(studyset):
-    """Convert nimads studyset object to a dataset."""
+def convert_nimads_to_dataset(studyset, annotation=None):
+    """Convert nimads studyset to a dataset."""
+    if isinstance(studyset, dict):
+        studyset = Studyset(studyset)
+    elif isinstance(studyset, str):
+        with open(studyset, "r") as f:
+            studyset = Studyset(json.load(f))
+    else:
+        raise ValueError(
+            "studyset must be: a dictionary, a path to a json file, or studyset object"
+        )
 
-    def _analysis_to_dict(study, analysis):
-        result = {
-            "metadata": {
-                "authors": study.name,
-                "journal": study.publication,
-                "title": study.name,
-                "sample_sizes": [study.metadata.get("sample_size")],
-            },
-            "coords": {
-                "space": analysis.points[0].space,
-                "x": [p.x for p in analysis.points],
-                "y": [p.y for p in analysis.points],
-                "z": [p.z for p in analysis.points],
-            },
-        }
-
-        if analysis.annotations:
-            result["labels"] = {}
-            for annotation in analysis.annotations.values():
-                result["labels"].update(annotation)
-
-        return result
-
-    def _study_to_dict(study):
-        return {
-            "metadata": {
-                "authors": study.authors,
-                "journal": study.publication,
-                "title": study.name,
-            },
-            "contrasts": {_create_name(a): _analysis_to_dict(study, a) for a in study.analyses},
-        }
-
-    return Dataset({_create_name(s): _study_to_dict(s) for s in list(studyset.studies)})
+    if annotation:
+        studyset.annotations = annotation
+    return nimads_to_dataset(studyset)
 
 
 def convert_neurosynth_to_dict(
